STATIC LIB Working fully.
Added all the necessary lines in .lua to link library to project.
The build directory for each lib and project is now different.
There is a even a folder for each kind (static or shared) for the future.
folders containing lib and project .cpp and .h files are now in their own folder (just for organisation)